(dev) documentation, minor code improvements
