+++ lib/haunted-pan-zoom.js
@@ -17,14 +17,19 @@ const HauntedPanZoom = ({ src, disabled, panStiffness, zoomStiffness }) => {
 		containerRef
 	} = usePanZoom(panStiffness, zoomStiffness);
 
+	// trigger load transition when src changes
 	useEffect(load, [src]);
+
+	// expose public api for interoperability
 	usePublicAPI(() => ({ zoomTo, resize }), [zoomTo, resize]);
 	usePublicAPI(() => ({ zoom }), [zoom]);
 
+	// don't render anything if src is missing
 	if (!src) {
 		return nothing;
 	}
 
+	// polymer interop
 	const dispatchEvent = useDispatchEvent();
 	useEffect(() => dispatchEvent(new CustomEvent('zoom-changed', {detail: {value: zoom}})), [dispatchEvent, zoom]);
 	useEffect(() => dispatchEvent(new CustomEvent('status-changed', {detail: {value: status}})), [dispatchEvent, status]);
