+++ lib/hooks/use-dispatch-event.js
@@ -1,5 +1,9 @@
 import { hook, Hook } from 'haunted';
 
+/**
+ * Returns the host element's dispatchEvent function.
+ * @return {Function} the dispatchEvent function
+ */
 export const useDispatchEvent = hook(class extends Hook {
 	constructor(id, el) {
 		super(id, el);
