+++ lib/hooks/use-pan.js
@@ -1,8 +1,13 @@
 import { useState, useEffect, useMemo } from 'haunted';
-import { getBounded } from './utils';
+import { getBounded } from '../utils';
 import { useMousePan, DRAGGING } from './use-mouse-pan';
 import { useTouchPan, DRAGGING as TOUCH_DRAGGING } from './use-touch-pan';
 
+/**
+ * Tracks mouse and touch movements and stores the panned distance relative to [0, 0].
+ * @param  {{x: Number[], y: Number[]}} bounds pan boundaries
+ * @return {Array} [panning: boolean, panX: number, panY: number, start: function, startTouch: function]
+ */
 export const usePan = (bounds = { x: [-100, 100], y: [-100, 100]}) => {
 	const [panX, setPanX] = useState(0);
 	const [panY, setPanY] = useState(0);
@@ -18,6 +23,7 @@ export const usePan = (bounds = { x: [-100, 100], y: [-100, 100]}) => {
 
 	const panning = status === DRAGGING || touchStatus === TOUCH_DRAGGING;
 
+	// when the user stops pinching, apply the bounds
 	useEffect(() => {
 		if (panning) {
 			return;
