+++ lib/hooks/use-pinch.js
@@ -1,42 +1,14 @@
 import { useEffect, useCallback, useReducer } from 'haunted';
+import { pinchReducer, initState, PINCHING } from '../reducers/pinch';
 
-export const INIT = 'init';
-export const PINCHING = 'pinching';
-
-const initState = {	status: INIT, lastDistance: 0, delta: 0 };
-
-const pinchReducer = (state, action) => {
-	switch (state.status) {
-
-	case INIT:
-		switch (action.type) {
-		case 'start':
-			return {
-				...state,
-				status: PINCHING,
-				lastDistance: action.initialDistance,
-				delta: 0
-			};
-		}
-		break;
-
-	case PINCHING:
-		switch (action.type) {
-		case 'stop':
-			return initState;
-		case 'pinch':
-			return {
-				...state,
-				delta: action.distance - state.lastDistance,
-				lastDistance: action.distance
-			};
-		}
-	}
-
-	return state;
-};
-
-export const usePinch = () => {
+/**
+ * Tracks the amount a user pinched the document.
+ *
+ * Tracking is started by calling the `start` function.
+ *
+ * @return {Array} [status: string, delta: number, start: function]
+ */
+const usePinch = () => {
 	const [state, dispatch] = useReducer(pinchReducer, initState);
 
 	// TODO: handle variable number of touches a la use-touch-pan
@@ -79,3 +51,5 @@ export const usePinch = () => {
 
 	return [state.status, state.delta, start];
 };
+
+export { usePinch, PINCHING };
