+++ lib/hooks/use-public-api.js
@@ -1,5 +1,11 @@
 import { hook, Hook, useMemo } from 'haunted';
 
+/**
+ * Adds public methods and values to the host element.
+ * @param {Function} fn     a function which returns a set.
+ * @param {Array}    values re-run the function when this value changes.
+ * @returns {void}
+ */
 export const usePublicAPI = hook(class extends Hook {
 	constructor(id, el, fn) {
 		super(id, el);
