+++ lib/hooks/use-resize.js
@@ -1,34 +1,31 @@
-import { useEffect, useRef } from 'haunted';
+import { useEffect } from 'haunted';
 
-export const useResize = (fn, ref) => {
+/**
+ * Reacts to size changes of an element.
+ * @param  {Function}    onResize function to call when the element is resized
+ * @param  {HTMLElement} el       a DOM element
+ * @return {void}
+ */
+export const useResize = (onResize, el) => {
 	if (!('ResizeObserver' in window)) {
 		return;
 	}
-	const fnRef = useRef();
 
+	// set up a new observer whenever the element or the callback changes
 	useEffect(() => {
-		fnRef.current = fn;
-	}, [fn]);
-
-	useEffect(() => {
-		const observer = new ResizeObserver(entries => {
-			const [{contentRect}] = entries;
-			if (!fnRef.current) {
-				return;
-			}
-			fnRef.current(contentRect);
-		});
-
-		if (ref.current) {
-			observer.observe(ref.current);
+		// do nothing if the element is missing
+		if (el == null) {
+			return;
 		}
 
-		if (fnRef.current && ref.current) {
-			fnRef.current(ref.current.getBoundingClientRect());
-		}
+		// set up the observer
+		const observer = new ResizeObserver(([{contentRect}]) => onResize(contentRect));
+		observer.observe(el);
 
-		return () => {
-			observer.unobserve(ref.current);
-		};
-	}, [ref.current]);
+		// trigger an initial resize so the size is updated when the referenced element changes
+		onResize(el.getBoundingClientRect());
+
+		// clean up the observer
+		return () => observer.unobserve(el);
+	}, [onResize, el]);
 };
