+++ lib/hooks/use-touch-pan.js
@@ -1,61 +1,17 @@
 import { useEffect, useCallback, useReducer } from 'haunted';
-import { center } from './utils';
-
-export const INIT = 'init';
-export const DRAGGING = 'dragging';
-
-const panReducer = (state, action) => {
-	switch (state.status) {
-
-	case INIT:
-		switch (action.type) {
-		case 'start':
-			return {
-				...state,
-				status: DRAGGING,
-				lastX: action.x,
-				lastY: action.y,
-				timeStamp: action.timeStamp
-			};
-		}
-		break;
-
-	case DRAGGING:
-		switch (action.type) {
-		case 'stop':
-			return {
-				...state,
-				status: INIT,
-				deltaX: 0,
-				deltaY: 0,
-				timeStamp: 0
-			};
-		case 'drag':
-			return {
-				...state,
-				deltaX: action.x - state.lastX,
-				deltaY: action.y - state.lastY,
-				lastX: action.x,
-				lastY: action.y,
-				timeStamp: action.timeStamp
-			};
-		case 'one-little-finger':
-			return {
-				...state,
-				lastX: action.x,
-				lastY: action.y
-			};
-		}
-	}
-
-	return state;
-};
-
-const initState = {	status: INIT, deltaX: 0, deltaY: 0, lastX: 0, lastY: 0, timeStamp: 0 };
+import { center } from '../utils';
+import { panReducer, initState, DRAGGING } from '../reducers/touch-pan';
 
 const centerFromEvent = ev => center(Array.from(ev.touches).map(t => [t.pageX, t.pageY]));
 
-export const useTouchPan = () => {
+/**
+ * Tracks touch movements on the document.
+ *
+ * Tracking is initialized by calling the start function and stops when touching ends.
+ *
+ * @return {Array} [status:string, deltaX: number, deltaY: number, timestamp: number, start: function]
+ */
+const useTouchPan = () => {
 	const [state, dispatch] = useReducer(panReducer, initState);
 
 	const start = useCallback(ev => {
@@ -64,12 +20,12 @@ export const useTouchPan = () => {
 		}
 
 		const [x, y] = centerFromEvent(ev);
-		dispatch({ type: 'start', x, y, timeStamp: ev.timeStamp });
+		dispatch({ type: 'start', x, y, timestamp: ev.timeStamp });
 	}, [dispatch]);
 
 	const onTouchMove = useCallback(ev => {
 		const [x, y] = centerFromEvent(ev);
-		dispatch({ type: 'drag', x, y, timeStamp: ev.timeStamp });
+		dispatch({ type: 'drag', x, y, timestamp: ev.timeStamp });
 	}, [dispatch]);
 
 	const onTouchEnd = useCallback(ev => {
@@ -102,5 +58,7 @@ export const useTouchPan = () => {
 		};
 	}, [state.status, onTouchStart, onTouchMove, onTouchEnd]);
 
-	return [state.status, state.deltaX, state.deltaY, state.timeStamp, start];
+	return [state.status, state.deltaX, state.deltaY, state.timestamp, start];
 };
+
+export { useTouchPan, DRAGGING };
